Move converting dashboard settings to dashboard command
Also move converting the settings data structure during saving into the
dashboard command.
